imp: category 구조를 db 에 저장 안하도록 변경

+++ src/main/kotlin/com/github/jyc228/musinsa/MusinsaException.kt
@@ -3,7 +3,7 @@ package com.github.jyc228.musinsa
 sealed class MusinsaException(message: String, cause: Throwable? = null) : RuntimeException(message, cause)
 
 class BrandNotFoundException(id: Long) : MusinsaException("brand $id not found")
-class CategoryNotFoundException(id: Long) : MusinsaException("category $id not found")
+class CategoryNotFoundException(id: Int) : MusinsaException("category $id not found")
 class ProductNotFoundException(id: Long) : MusinsaException("product $id not found")
 
 class InvalidRequestException(val param: String, message: String) : MusinsaException(message)
